moved provider to folder, injected _fileSystem and _httpClient
